some initial thoughts on the organization of this project
Rather than making the choice of role *explicit*, as in the existing
example, this makes the puppet role *implict* based on the ansible role
selected.  In this model, your top-level playbook could have:

    roles:
      - profile/controller

Or:

    roles:
      - profile/compute

And so forth to select between different host roles. Dependencies (e.g.,
for the `common` role that is required everywhere) are handled via
Ansible's role metadata support (`roles/<rolename>/meta/main.yml`)

It's also nice to make things reasonably granular...e.g., have roles
that *require* puppet depend on a role that is responsible for install
it.  This just helps keep playbooks focused.

With these changes, the example deployment playbook becomes:

    - hosts: controllers
      roles:
        - profile/controller
        - puppet_run